make LIFOs and pool fully thread-safe
The lock-less implementation of LIFOs now has no limitation on the numbers
of producer and consumer threads. The only assumption is that we need to know
the maximum length of the LIFO on initialization, and we will not be able
to grow it afterwards. Therefore the ulifo API became completely identical to
upool API, and upool semantics were removed, all calls to upool renamed.

The under-the-hood implementation uses a new data structure called uring,
which allows pointer tagging to avoid the ABA problem.