rewrite uprobe_dejitter
The goal is to comply with ISO 13818-1, which specifies strong limits on the
variations of the STC (cr_prog) clock. A phase-locked loop has been added
after the low-pass filter to smooth out variations, and the variations of
the frequency of the PLL are also smoothed out. The new algorithm now
requires libm.

The API of uprobe_dejitter_alloc and uprobe_dejitter_set was modified. The last
argument is now a boolean (enabled or not) instead of a constant for the
low-pass filter, and a new argument after this one specifies the initial
standard deviation. By default (set to 0) it is 15 ms. For low latency
use cases, set to 1.