upipe_mpgaf: fix potential assert
Reset flow_def_requested when a new flow format is provided as it is
used to check if the pipe can handle new urefs. Because the provided
flow format will unset the ubuf manager before requiring a new one, the
pipe must wait for the new ubuf manager before handling new urefs.