upipe_mpgaf: set default octet rate to 8kb/s
Set the octet rate default value to 8kb/s and use it for flow format
negotiation.
This value will be used in the flow def before the first outputted
packet as the framer will force a need output probe event to negotiate
the output flow format.
At this time the real octet rate is unknown for AAC streams because the
flow format is not negotiated but some pipes need it in the flow def to
accept it.
These pipes may not accept the octet rate to decrease afterwards so it
must be as low as possible.
For AAC streams, octet rates are rounded up to a multiple of 8kb/s so
the lowest value possible is 8kb/s.
For MPEG layer 1, 2 and 3 streams, octet rates are parsed before the
probe event, so the default value will be overwritten.