upipe_freetype: fix potential SEGV
Replace FTC_Manager_LookupFace with FTC_Manager_LookupSize to make sure
face size is not NULL and will not be.
It seems that FTC_Manager_LookupFace does not guarantee that face->size
will not be reset to NULL at any time but FTC_Manager_LookupSize does.
Also make sure we remove the FaceID (font pointer) from the manager
before we free it as recommended by FreeType documentation.