srth: rework unexpected handshake packets handling
Ignore handshakes with different socket id when we're already connected.
Do not send a reply, as it would reach the wrong remote.
keepalive will trigger and restart after 10s if needed

Second case: if we do not received the expected induction type, trigger disconnection.
Since we sent a rejection message we expect a new socket anyway.